pyproject.toml
[project]
name = "mlk-appointments-chatbot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.123.4",
    "uvicorn>=0.38.0",
    "sqlmodel==0.0.19",
    "python-dotenv==1.0.1",
    "httpx==0.27.0",
    "dateparser==1.2.0",
    "psycopg2-binary>=2.9.9",
    "pydantic[email]>=2.0.0"
]
